Issue #12: move factory method to the init python file
