+++ scripts/autoscaling/node_reassign.py
@@ -14,18 +14,8 @@
 
 import argparse
 
+import pipeline.autoscaling as autoscaling
 from pipeline.autoscaling import *
-
-
-def create_cloud_provider(cloud, cloud_region):
-    if cloud == "aws":
-        return awsprovider.AWSInstanceProvider(cloud_region)
-    elif cloud == "az":
-        return azureprovider.AzureInstanceProvider(cloud_region)
-    elif cloud == "gcloud":
-        return gcpprovider.GCPInstanceProvider(cloud_region)
-    else:
-        raise RuntimeError("Cloud: {} is not supported".format(cloud))
 
 
 def main():
@@ -40,7 +30,7 @@ def main():
 
     kube_provider = kubeprovider.KubeProvider()
     cloud_region = kube_provider.get_cloud_region(old_id)
-    cloud_provider = create_cloud_provider(cloud, cloud_region)
+    cloud_provider = autoscaling.create_cloud_provider(cloud, cloud_region)
 
     ins_id = cloud_provider.find_and_tag_instance(old_id, new_id)
     nodename, nodename_full = cloud_provider.get_instance_names(ins_id)
