+++ scripts/autoscaling/nodedown.py
@@ -14,19 +14,8 @@
 
 import argparse
 
+import pipeline.autoscaling as autoscaling
 from pipeline.autoscaling import *
-
-
-def create_cloud_provider(cloud, cloud_region):
-    if cloud == "aws":
-        return awsprovider.AWSInstanceProvider(cloud_region)
-    elif cloud == "az":
-        return azureprovider.AzureInstanceProvider(cloud_region)
-    elif cloud == "gcloud":
-        return gcpprovider.GCPInstanceProvider(cloud_region)
-    else:
-        raise RuntimeError("Cloud: {} is not supported".format(cloud))
-
 
 def main():
     parser = argparse.ArgumentParser()
@@ -38,7 +27,7 @@ def main():
 
     kube_provider = kubeprovider.KubeProvider()
     cloud_region = kube_provider.get_cloud_region(run_id)
-    cloud_provider = create_cloud_provider(cloud, cloud_region)
+    cloud_provider = autoscaling.create_cloud_provider(cloud, cloud_region)
 
     try:
         ins_id = cloud_provider.find_instance(run_id)
